Changed interactable indicator to particle system
