--- Assets/Animations/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 289dc20e7a794344cad569d026a527c8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
