+++ Assets/Materials/Effects/ParticleSave.mat
@@ -116,7 +116,7 @@ Material:
     m_Colors:
     - _BaseColor: {r: 1, g: 1, b: 1, a: 1}
     - _Color: {r: 1, g: 0.9027963, b: 0, a: 1}
-    - _Emission: {r: 1.5365012, g: 0.3016538, b: 0.16524625, a: 1}
+    - _Emission: {r: 1.5365013, g: 0.78304183, b: 0.16893467, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 1}
   m_BuildTextureStacks: []
