+++ Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
@@ -27,6 +27,33 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 60000
+--- !u!114 &-8577785680029822455
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/Character/Enemy/EyeballAttack
+  m_EditorClassIdentifier: 
+  Path: event:/Character/Enemy/EyeballAttack
+  Guid:
+    Data1: -807238135
+    Data2: 1150889072
+    Data3: -1722838893
+    Data4: -11051897
+  Banks:
+  - {fileID: 2073874680224088596}
+  IsStream: 0
+  Is3D: 0
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 1272
 --- !u!114 &-8252471984227306234
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -201,10 +228,10 @@ MonoBehaviour:
   Path: NoireFM/Build/Desktop/Master.strings.bank
   Name: Master.strings
   StudioPath: bank:/Master.strings
-  lastModified: 638347954939734242
+  lastModified: 638348307575105430
   FileSizes:
   - Name: Desktop
-    Value: 1624
+    Value: 1818
   Exists: 1
 --- !u!114 &-5288706457553147157
 MonoBehaviour:
@@ -221,7 +248,7 @@ MonoBehaviour:
   Path: NoireFM/Build/Desktop/OST.bank
   Name: OST
   StudioPath: bank:/OST
-  lastModified: 638347956491612264
+  lastModified: 638348307575634594
   FileSizes:
   - Name: Desktop
     Value: 7228224
@@ -297,6 +324,33 @@ MonoBehaviour:
   - Name: Desktop
     Value: 3609344
   Exists: 1
+--- !u!114 &-3670064223993183543
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/MainMenu/Select
+  m_EditorClassIdentifier: 
+  Path: event:/MainMenu/Select
+  Guid:
+    Data1: -1917602514
+    Data2: 1082209155
+    Data3: 600207258
+    Data4: -655676877
+  Banks:
+  - {fileID: 2073874680224088596}
+  IsStream: 0
+  Is3D: 0
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 1000
 --- !u!114 &-2952506944820352129
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -525,6 +579,10 @@ MonoBehaviour:
   - {fileID: -2952506944820352129}
   - {fileID: -8154400465538517080}
   - {fileID: 440793489035882025}
+  - {fileID: -8577785680029822455}
+  - {fileID: -3670064223993183543}
+  - {fileID: 1770758193841697651}
+  - {fileID: 327791649936647755}
   EditorParameters:
   - {fileID: 4793595672983564214}
   - {fileID: 4320370547615308145}
@@ -534,8 +592,35 @@ MonoBehaviour:
   - {fileID: 5191857318038855451}
   StringsBanks:
   - {fileID: -5436910303919886702}
-  cacheTime: 638347956491612264
+  cacheTime: 638348307575794825
   cacheVersion: 131607
+--- !u!114 &327791649936647755
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/Character/Enemy/EyeballDie
+  m_EditorClassIdentifier: 
+  Path: event:/Character/Enemy/EyeballDie
+  Guid:
+    Data1: 940624867
+    Data2: 1131460716
+    Data3: -753849419
+    Data4: 1841561817
+  Banks:
+  - {fileID: 2073874680224088596}
+  IsStream: 0
+  Is3D: 0
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 3374
 --- !u!114 &440793489035882025
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -619,6 +704,33 @@ MonoBehaviour:
   MinDistance: 1
   MaxDistance: 20
   Length: 216058
+--- !u!114 &1770758193841697651
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ceb653cd98e289a4e8697a1af55201f2, type: 3}
+  m_Name: event:/MainMenu/Click
+  m_EditorClassIdentifier: 
+  Path: event:/MainMenu/Click
+  Guid:
+    Data1: -47322736
+    Data2: 1312629771
+    Data3: -742009702
+    Data4: -219181940
+  Banks:
+  - {fileID: 2073874680224088596}
+  IsStream: 0
+  Is3D: 0
+  IsOneShot: 1
+  Parameters: []
+  MinDistance: 1
+  MaxDistance: 20
+  Length: 1447
 --- !u!114 &1987584681221027207
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -661,10 +773,10 @@ MonoBehaviour:
   Path: NoireFM/Build/Desktop/SFX.bank
   Name: SFX
   StudioPath: bank:/SFX
-  lastModified: 638337975374717147
+  lastModified: 638348307575794825
   FileSizes:
   - Name: Desktop
-    Value: 3612864
+    Value: 3697024
   Exists: 1
 --- !u!114 &2725092569944637218
 MonoBehaviour:
@@ -905,7 +1017,7 @@ MonoBehaviour:
   Path: NoireFM/Build/Desktop/Master.bank
   Name: Master
   StudioPath: bank:/Master
-  lastModified: 638347954939723169
+  lastModified: 638343696432621779
   FileSizes:
   - Name: Desktop
     Value: 2314
