+++ Assets/Prefabs/Environment/Door.prefab
@@ -219,7 +219,7 @@ PrefabInstance:
     - target: {fileID: 4568057972659828093, guid: 8325d3eece07f6940804a262516bedc2,
         type: 3}
       propertyPath: m_Name
-      value: InteractableIndicator (3)
+      value: InteractableIndicator
       objectReference: {fileID: 0}
     - target: {fileID: 4568057972659828093, guid: 8325d3eece07f6940804a262516bedc2,
         type: 3}
@@ -231,12 +231,36 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 8325d3eece07f6940804a262516bedc2, type: 3}
---- !u!224 &2466484144748518715 stripped
-RectTransform:
+--- !u!4 &2466484144748518715 stripped
+Transform:
   m_CorrespondingSourceObject: {fileID: 3575871191110788994, guid: 8325d3eece07f6940804a262516bedc2,
     type: 3}
   m_PrefabInstance: {fileID: 1412649953652144825}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &8555542048314514461 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 7287344867881789092, guid: 8325d3eece07f6940804a262516bedc2,
+    type: 3}
+  m_PrefabInstance: {fileID: 1412649953652144825}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 98448373a9a7c114ca2c0f2ce9ee08bf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &9103546960607671571 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 7911969261213650858, guid: 8325d3eece07f6940804a262516bedc2,
+    type: 3}
+  m_PrefabInstance: {fileID: 1412649953652144825}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 98448373a9a7c114ca2c0f2ce9ee08bf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &3009936685229568296
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -248,12 +272,12 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_RootOrder
-      value: -1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalScale.x
-      value: 199.99998
+      value: 200.00003
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
@@ -263,27 +287,27 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalScale.z
-      value: 199.99998
+      value: 200.00003
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: 93.19
+      value: -14.5
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: 11.29
+      value: 4.66
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: 114.14
+      value: -30.00693
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalRotation.w
-      value: 0.69968057
+      value: 0.7071068
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
@@ -293,7 +317,7 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalRotation.y
-      value: -0.71445584
+      value: 0.7071068
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
@@ -308,7 +332,7 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
-      value: -91.197
+      value: 90
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
         type: 3}
@@ -456,10 +480,12 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   ID: door
-  interactText: 
-  maxInteractions: 0
-  interactableIndicator: {fileID: 0}
-  destinationScene: 1
+  interactText: Enter
+  maxInteractions: 1
+  interactableIndicator: {fileID: 9103546960607671571}
+  onInteractIndicator: {fileID: 8555542048314514461}
+  destinationScene: 4
+  doorAnimator: {fileID: 4489102591789011697}
 --- !u!4 &3337105797287712451 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: -8679921383154817045, guid: e03f38cf1728ffe4481761dcc761ddea,
