+++ Assets/Prefabs/Environment/TakeoutBoxes.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 0d5a8221b3f4d5a48b08c01aa3016c2b
-folderAsset: yes
-DefaultImporter:
+guid: 84535f0f9beaa954b9f60747f9a08918
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
