--- Assets/Scenes/BedrockPlains/NavMesh-Plane.133 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1d609c0455775dc44b0c6ce7645bdb33
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
