--- Assets/Scenes/BedrockPlains/NavMesh-Plane.133.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6e468040a568fe546a202412d2ef0b7c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
