+++ Assets/Scripts/BedRockPlains/Lamp.cs
@@ -1,10 +1,8 @@
-﻿using UnityEngine;
-
-public class Lamp : InteractableObject
+﻿public class Lamp : InteractableObject
 {
-    [SerializeField] private BGMAudio bgmAudio;
     public override void Interact()
     {
+        onInteractIndicator.Play();
         interactionsOccured++;
         GameEventsManager.Instance.BedrockPlainsEvents.LampInteract();
         FinishInteract();
