+++ Assets/Scripts/Gameplay/Interaction/InteractableDialogue.cs
@@ -6,6 +6,7 @@ public class InteractableDialogue : InteractableObject
     [SerializeField] private TextAsset inkJSONAsset;
     public override void Interact()
     {
+        onInteractIndicator.Play();
         interactionsOccured++;
         var story = new Story(inkJSONAsset.text);
         DialogueUI.Instance.StartDialogue(story);
