+++ Assets/Scripts/Gameplay/Interaction/InteractableExamine.cs
@@ -6,6 +6,7 @@ public class InteractableExamine : InteractableObject
     [TextArea(3, 5)] [SerializeField] private string examineText;
     public override void Interact()
     {
+        onInteractIndicator.Play();
         interactionsOccured++;
         ExamineUI.Instance.Display(examineText);
         FinishInteract();
